% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoxx.R, R/timeconversion.R
\name{isoyear_c}
\alias{isoyear_c}
\title{ISO year (character) from Date object}
\usage{
isoyear_c(x = lubridate::today())

isoyear_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year in character

ISO year in character
}
\description{
ISO year (character) from Date object

ISO year (character) from Date object
}
\examples{
isoyear_c('2021-08-11')
isoyear_c(lubridate::today())
isoyear_c('2021-08-11')
isoyear_c(lubridate::today())
}
