% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoxx.R, R/timeconversion.R
\name{isoyearweek_c}
\alias{isoyearweek_c}
\title{ISO year and week (character) from Date object}
\usage{
isoyearweek_c(x = lubridate::today())

isoyearweek_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year and week in character

ISO year and week in character
}
\description{
ISO year and week (character) from Date object

ISO year and week (character) from Date object
}
\examples{
isoyear_c('2021-08-11')
isoyear_c(lubridate::today())
isoyear_c('2021-08-11')
isoyear_c(lubridate::today())
}
