% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearweek_split.R
\name{isoyearweek_to_week_n}
\alias{isoyearweek_to_week_n}
\title{ISO yearweek to week (numeric)}
\usage{
isoyearweek_to_week_n(yrwk)
}
\arguments{
\item{yrwk}{Year-week, e.g. "2020-19" for 19th week in 2020}
}
\description{
This function breaks the string connected with '-' into year/week
}
